guildroom/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
django-cors-headers==2.2.0
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
pafy==0.5.4
Pillow==5.0.0
pip==9.0.1
python-vlc==3.0.102
pytz==2018.3
requests==2.18.4
setuptools==38.5.2
six==1.11.0
urllib3==1.22
vine==1.1.4
wheel==0.30.0
youtube-dl==2018.3.10
django-cleanup==2.1.0
